requirements.txt
pytorch
numpy
scipy